+++ lib/src/main/java/com/smileidentity/models/PrepUpload.kt
@@ -6,6 +6,7 @@ import com.smileidentity.BuildConfig
 import com.smileidentity.SmileID
 import com.smileidentity.models.v2.Metadatum
 import com.smileidentity.networking.calculateSignature
+import com.smileidentity.util.getCurrentIsoTimestamp
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
@@ -21,7 +22,7 @@ data class PrepUploadRequest(
     @Json(name = "retry") val retry: Boolean? = null,
     @Json(name = "source_sdk") val sourceSdk: String = "android",
     @Json(name = "source_sdk_version") val sourceSdkVersion: String = BuildConfig.VERSION_NAME,
-    @Json(name = "timestamp") val timestamp: String = System.currentTimeMillis().toString(),
+    @Json(name = "timestamp") val timestamp: String = getCurrentIsoTimestamp(),
     @Json(name = "signature") val signature: String = calculateSignature(timestamp),
 )