+++ lib/src/main/java/com/smileidentity/models/ProductsConfig.kt
@@ -3,6 +3,7 @@ package com.smileidentity.models
 import android.os.Parcelable
 import com.smileidentity.SmileID
 import com.smileidentity.networking.calculateSignature
+import com.smileidentity.util.getCurrentIsoTimestamp
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import kotlinx.parcelize.Parcelize
@@ -10,7 +11,7 @@ import kotlinx.parcelize.Parcelize
 @JsonClass(generateAdapter = true)
 data class ProductsConfigRequest(
     @Json(name = "partner_id") val partnerId: String = SmileID.config.partnerId,
-    @Json(name = "timestamp") val timestamp: String = System.currentTimeMillis().toString(),
+    @Json(name = "timestamp") val timestamp: String = getCurrentIsoTimestamp(),
     @Json(name = "signature") val signature: String = calculateSignature(timestamp),
 )