+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Release Notes
+## Unreleased
+
+### Changed
+* Timestamp consistency from date epoch to iso format 
 
 ## 10.4.3 (Unreleased)